Scene 4 a And b fixed
The script had scene 9a, 8a etc. In the next scene button functions. which don't exist, it would just be scene 9 or 8 etc.
Also 4a had and extra script component in the canvas inspector which had nothing in it so I got rid of it. The first script component already had everything and works fine.
Everything should work, it was fine on my end.